# Upgrade to go 1.24