--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/grafana/tempo
 
-go 1.23.3
+go 1.24.0
 
 require (
 	cloud.google.com/go/storage v1.41.0