@joe-elliott 2025-02-12
This PR requires us to update our tools image first which requires a go 1.24 base image. This does not exist yet so we will hold for a bit.