@ruslan-mikhailov 2025-02-14
I continue the PR here: https://github.com/grafana/tempo/pull/4704